# openapi.yaml
openapi: 3.0.3
info:
  title: Ricerca Parcheggio Efficace - OpenAPI 3.0
  description:
    Questa API consente di interagire con il servizio Ricerca Parcheggio Efficace per sfruttare le sue funzionalità.
  version: 0.0.1
paths:
  /v1/parcheggio:
    post:
      summary: Aggiunta parcheggio
      description: Viene aggiunto un nuovo parcheggio con i campi inviati
      operationId: postParcheggio
      tags: 
        - parcheggi
      security:
        - BearerAuth: [admin]
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - oneOf:
                  - $ref: '#/components/schemas/ParcheggioFreeDiscoPostReqSchema'
                  - $ref: '#/components/schemas/ParcheggioFreePostReqSchema' 
                  discriminator:
                    propertyName: isDisco
                - $ref: '#/components/schemas/ParcheggioPayReqSchema'
                - $ref: '#/components/schemas/ParcheggioVigilatoReqSchema'
              discriminator:
                propertyName: _type
            examples:
                free:
                  summary: esempio parcheggio gratuito senza disco orario
                  value: {
                          "_type": "ParcheggioFree",
                          "nome": "p1",
                          "posizione":{
                            "type": "Point",
                            "coordinates": [11, 14]
                          },
                          "numPosti": 23,
                          "isCoperto": true,
                          "statoParcheggio": "Non disponibile",
                          "numPostiDisabili": 2,
                          "numPostiGravidanza": 0,
                          "numPostiAuto": 15,
                          "numPostiMoto": 6,
                          "numPostiFurgone": 0,
                          "numPostiBus": 0,
                          "isDisco": false
                        }
                freeDisco:
                  summary: esempio parcheggio gratuito con disco orario
                  value: {
                          "_type": "ParcheggioFree",
                          "nome": "p2",
                          "posizione": {
                            "type": "Point",
                            "coordinates": [9, 4]
                          },
                          "numPosti": 23,
                          "isCoperto": false,
                          "statoParcheggio": "Non disponibile",
                          "numPostiDisabili": 2,
                          "numPostiGravidanza": 0,
                          "numPostiAuto": 15,
                          "numPostiMoto": 6,
                          "numPostiFurgone": 0,
                          "numPostiBus": 0,
                          "isDisco": true,
                          "dataInizio": "13:35",
                          "dataFine": "16:35"
                        }
                pay:
                  summary: esempio parcheggio a pagamento
                  value: {
                          "_type": "ParcheggioPay",
                          "nome": "p3",
                          "posizione":{
                            "type": "Point",
                            "coordinates": [24, 1]
                          },
                          "numPosti": 23,
                          "isCoperto": false,
                          "statoParcheggio": "Non disponibile",
                          "numPostiDisabili": 2,
                          "numPostiGravidanza": 1,
                          "numPostiAuto": 13,
                          "numPostiMoto": 6,
                          "numPostiFurgone": 1,
                          "numPostiBus": 0,
                          "tariffa": 1.20
                        }
                vigilato:
                  summary: esempio parcheggio vigilato
                  value: {
                        "_type": "ParcheggioVigilato",
                        "nome": "p4",
                        "posizione":{
                          "type": "Point",
                          "coordinates": [0, 0]
                        },
                        "numPosti": 30,
                        "isCoperto": true,
                        "statoParcheggio": "Non disponibile",
                        "numPostiDisabili": 2,
                        "numPostiGravidanza": 0,
                        "numPostiAuto": 15,
                        "numPostiMoto": 6,
                        "numPostiFurgone": 0,
                        "numPostiBus": 7,
                        "postiOccupati": {
                          "postiOcc": 10,
                          "postiOccDisabili": 1,
                          "postiOccGravidanza": 0,
                          "postiOccAuto": 6,
                          "postiOccMoto": 2,
                          "postiOccFurgone": 0,
                          "postiOccBus": 1
                        },
                        "tariffa": 2.50
                      }
      responses:
        '201':
          $ref: "#/components/responses/PostSuccessfulRes"
        '5XX': 
          $ref: "#/components/responses/GeneralError"
    get:
      summary: Richiesta lista parcheggi
      description: Ritorna la lista di tutti i parcheggi
      operationId: getAllParcheggi
      tags: 
        - parcheggi
      responses:
        '200':
          $ref: '#/components/responses/GetAllSuccessfulRes'
        '4XX': 
          $ref: "#/components/responses/GeneralError"
        '5XX': 
          $ref: "#/components/responses/GeneralError"
...
components:
...
  schemas:
...
    ParcheggioSchema:
      type: object
      required:
        - _type
        - nome
        - posizione
        - numPosti
        - isCoperto
        - statoParcheggio
        - numPostiDisabili
        - numPostiGravidanza
        - numPostiAuto
        - numPostiMoto
        - numPostiFurgone
        - numPostiBus
      properties:
        _type:
          type: string
        nome:
          type: string
        posizione:
          type: object
          properties:
            type:
              type: string
              enum:
              - Point
            coordinates:
              type: array
              minItems: 2
              maxItems: 2
              items:
                type: number
        numPosti:  
          type: integer
        isCoperto:
          type: boolean
        statoParcheggio:
          type: string
          enum:
          - Disponibile
          - Non disponibile
        numPostiDisabili: 
          type: integer
        numPostiGravidanza:
          type: integer
        numPostiAuto:
          type: integer
        numPostiMoto: 
          type: integer
        numPostiFurgone:
          type: integer
        numPostiBus:
          type: integer
...
    ParcheggioFreeDiscoPostReqSchema:
      allOf:
        - $ref: '#/components/schemas/ParcheggioSchema'
        - type: object
          required:
            - isDisco
            - dataInizio
            - dataFine
          properties:
            isDisco:
              type: boolean
              enum:
              - true
            dataInizio:
              type: string
            dataFine:
              type: string
    ParcheggioFreePostReqSchema:
      allOf:
        - $ref: '#/components/schemas/ParcheggioSchema'
        - type: object
          required:
            - isDisco
          properties:
            isDisco:
              type: boolean
              enum:
              - false
    ParcheggioPayReqSchema:
      allOf:
        - $ref: '#/components/schemas/ParcheggioSchema'
        - type: object
          required:
            - tariffa
          properties:
            tariffa:
              type: number
    ParcheggioVigilatoReqSchema:
      allOf:
        - $ref: '#/components/schemas/ParcheggioSchema'
        - type: object
          required:
            - postiOccupati
            - tariffa
          properties:
            postiOccupati:
              type: object
              required:
                - postiOcc
                - postiOccDisabili
                - postiOccGravidanza
                - postiOccAuto
                - postiOccMoto
                - postiOccFurgone
                - postiOccBus
              properties:
                postiOcc:
                  type: integer
                postiOccDisabili:
                  type: integer
                postiOccGravidanza:
                  type: integer
                postiOccAuto:
                  type: integer
                postiOccMoto:
                  type: integer
                postiOccFurgone:
                  type: integer
                postiOccBus:
                  type: integer
            tariffa:
              type: number
...
    ParcheggioShortResSchema:
      type: object
      required:
        - _id
        - _type
        - nome
        - request
      properties:
        _id:
          type: string
        _type:
          type: string
        nome:
          type: string
        request:
          type: object
          required:
            - type
            - url
          properties:
            type:
              type: string
              enum:
              - GET
            url:
              type: string
...
    ParcheggioPostResSchema:
      type: object
      required:
        - message
        - createdParcheggio
      properties:
        message:
          type: string
        createdParcheggio:
          $ref: '#/components/schemas/ParcheggioShortResSchema'
...
    GetParcheggiResSchema:
      type: object
      required: 
       - count
       - parcheggi
      properties:
        count:
          type: integer
        parcheggi:
          type: array
          items:
            $ref: '#/components/schemas/ParcheggioShortResSchema'
...
    GeneralErrorSchema:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 
  responses:
...
    GeneralError:
      description: "Errore generico per cui viene mostrato un messaggio di errore"
      content: 
        application/json:
          schema:
            $ref: "#/components/schemas/GeneralErrorSchema"
          example: 
            error: "Error"
    PostSuccessfulRes:
      description: "Risposta alla POST in caso di successo"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ParcheggioPostResSchema' 
      links:
        GetParcheggioById: 
          $ref: '#/components/links/GetParcheggioById'
...
    GetAllSuccessfulRes:
      description: Risposta in caso di successo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetParcheggiResSchema'
      links:
        GetParcheggioById: 
          $ref: '#/components/links/GetParcheggioById'
...
  links:
    GetParcheggioById:
      operationId: getParcheggioById
      parameters:
        parcheggioId: 'response.body#/_id'
      description: Il valore '_id' ritornato può essere usato come 'parcheggioId' in 'GET /parcheggi/{parcheggioId}'